Improve registration notification email messaging
If multiple workshops are running simultaneously, it's pretty difficult
to figure out what workshop (and where) the student signed up for, e.g.
two Test-Driven Rails workshops are running back-to-back, one in SF and
the other in Boston. This improves email messaging to include the city
the workshop is running in, as well as the dates.