Consitent layout for purchases#watch & videos#show
Test coverage for multi-video purchases.

Refactor purchases#watch to redirect.

purchases#watch had two responsibilities: It could display a list of videos
or a single video depending on the product.

Since there was already a nested videos resource with a show action, it
reduces the logic and repitition in the view layer to redirect there in
the single video case and to videos#index in the multiple video case.

Prevent unpaid access to videos.

Protection already existed in the purchases#watch, but not on individual
video pages.

apply new helpers

fix double render problem