Turn off product pricing A/B test
This avoids skewing results. We're not providing alternate product prices, so it
will just fall back on the original pricing. Since the "alternate" pricing
bucket will now have results for when the user saw both alternate and primary
pricing, the results will be skewed.

Better to just turn it off for now.