Add Google Webmaster Tools verification file
This file is used as verification of our ownership of
learn.thoughtbot.com. I've tried the TXT and CNAME methods of
verification from DNSimple but those DNS records are not being
recognized.

The purpose of setting up Google Webmaster Tools is to improve our
Search Engine Optimization data such as:

* External and internal links to the site.
* Search keywords and their impressions, clicks, CTR, and average
  position.
* Author stats.

https://www.apptrajectory.com/thoughtbot/learn/stories/15627125