schedule cancelations for end of billing period
This fulfills this story:
https://www.apptrajectory.com/thoughtbot/learn/stories/15627152

This makes the change so that when a user cancel's the subscription, we
tell stripe about the cancelation and tell stripe that we want it to
continue for the remainder of the billing period. We then note the last
date of the period on our subscription.

When the final date of cancelation comes, Stripe actually deletes the
subscription and sends us a webhook about it. We subscribe to this
webhook and actually deactivate the subscription, also removing access
to all the products and workshops they purchased under the subscription,
and also send them the subscription cancelation follow up email.