Use same default values in :with_github user trait as in mocked auth
We have specific values set for the email address and Github username that is
returned by the mocked Omniauth response. To ensure we're testing against the
same user object that connected to a GitHub account, we should set the
github-related attributes in the user factory to the same values as
what we're getting back from Omniauth.