We need to be able to make mentors unavailable
I realized this morning that Ben and Myself had been set to
`available_to_mentor: false`, and we were not carried over to the new
Mentor records because of that.

However, there are still paying customers who have us as their mentor,
and someone who has us would not get an error because we had no mentor
record.

So this add a boolean to to indicate whether a Mentor is available for
new mentees and uses it in the right places.