Update spec to remove control coupling; fix model
While I was working on fixing the control coupling the tests started to
show failures in the work that I had done to clean up the Demeter
issues. This commit cleans up the coupling in the tests and gets the new
model methods to do as they were supposed to before.