Allow customer's to refund themselves
This new addition gives a cancelling customer the option to get a refund
after they cancel.

If the user tries to go back to this page after they've already gotten a
refund, and tries to do another one, they will get a 500 error. Stripe
won't allow you to refund the same charge twice. I figure this is
acceptable behavior for now.