Show preview video instead of thumbnail if present
https://www.apptrajectory.com/thoughtbot/learn/stories/15641131

When viewing a public video page, show preview video instead of
thumbnail when one is present.

Use a the request's host rather than wistia for video urls in tests.
This allows the url to be output to the page without actually hitting
the wistia api. Some tests explicitly check this url for the video hash
to ensure that the promo and not the real video was shown.