Extract `Clip` from `Video`
The ActiveRecord `Video` class handled multiple responsibilities including
the concept of an embeddable video and the concept of something a little
higher level that encapsulated a video, notes, a promo video, and a
thumbnail. The main video and promo video responsibilities were a
particular concern because the methods to deal with them were duplicated
with the prefix `preview_`.

The concept of an embeddable video was extracted into a PORO named
`Clip`. Both the main video clip and the promo clip are `Clip` objects.

`Video#preview` will return an `VideoThumbnail` object if there is no
preview. This is a light wrapper around a url to a video thumbnail. We
take advantage of `to_partial_path` in the view, eliminating the
conditional there.