redirect back to return_to after auth with github
Fixes: https://www.apptrajectory.com/thoughtbot/learn/stories/15632156

We upgraded clearance to get a fix for query parameters being preserved
after a redirect.

Upgrading clearance to the latest version broke a few things, which I
was able to fix by upgrading bourbon (which resulted in downgrading
sass).

The new version of clearance doesn't work the same way for translations,
so I hardcoded that path. We have coverage that checks for the right
path being in the message, so I feel comfortable with having it
hardcoded in the translation.