Addresses OAuth redirection bug
When upcase (as provider) authenticates a user, it redirects back to
`session[:return_to]`, set to the authorize path that the client application
triggered.

Chad suggested to clear the session variable upon successful authentication, but
for some reason it wasn't being cleaned.

Problem was that we set that session value always on
`resource_owner_authenticator` -which is akin to a doorkeeper `before_filter`-
when we only want to remember where to redirect when an unauthenticated user
gets to the sign in form in Upcase (ant not when it's already authenticated).

The flow before was:

1. client asks upcase for authorization. Doorkeeper sets the session variable.
2. In upcase, user clicks on "sign in with Github", triggering a successful
   OAuth flow. As of Chad's change, session is cleared after next step.
3. Redirect back to `return_to`, which was the authorize link. Complete a
   successful OAuth flow between client and Upcase as provider.
4. Strategy in client asks for current_user information (`api/v1/me.json`), this
   code runs on Upcase, and before this changeset it was setting
   `session[:return_to]` _again_. So any future authentication in upcase would
   again trigger this same OAuth flow between client and upcase.

Now, doorkeeper only sets `session[:return_to]` only when unauthenticated, and
the `clear_return_to` then works.