Fix broken plans page post-repo clone
The bin/setup script was creating a plans page that 500'ed due to the
`upcase-249` slug not matching the expected `prime-249` slug. After fixing that,
I noticed we were creating 8 plans in the bin/setup script. None of them exactly
matched what we have for plans in production. So, I fixed that while in there.

I'll attach before and after screenshots to the pull request.