Prevent exception when pressing return on checkout
When user presses [Return] key on the form, it triggers `click` event on
the nearest submit button, which in this case is the "Apply coupon"
submit button. We also bind the `click` event on that button to check
the validity of the coupon code. However, it turned out that we did not
check if the coupon code is there first before trying to making a
request, resulting in sending a request to non-existence endpoint.

This commit includes the changes that:

* Trim the coupon code that user enters.
* Prevent coupon code validation if the coupon code is empty.
* Change the rescue class on Coupon model, as Stripe now raises
  `InvalidRequestError` instead of `APIError` on non-existence coupon.
* Change coupon submit button to normal button, so it makes the [Return]
  key triggers the "Submit Payment" button, which is expected.

https://trello.com/c/j95ImHDc
https://thoughtbot.airbrake.io/projects/6325/groups/1251484741652927019
https://thoughtbot.airbrake.io/projects/6325/groups/1240487883421998395