"Next Up" should be shown as "Exercise"
First exercise of an unstarted trail should have the `NEXT_UP` state. But in
other views (like topics) we want to show the proper title, "Exercise", instead
of "Next Up". This changeset marks the first exercise as next, but changes the
label to be "exercise".

It is funky because we are reusing the query object to cache statuses not only
in trails, but also in topics.

https://trello.com/c/MgPjT8Dp/218-optimize-status-for-exercises-within-trails