Allow exercises reordering in trail admin page
This is done by enabling orderable in rails_admin edit page, and
override `exercise_ids=` setter for Trail to update the position.

Note that to keep the code simple it means that we will generate N more
queries after updating exercises in trail, where N is the number of
exercises in trail, but performance impact should not be that much.

https://github.com/sferik/rails_admin/wiki/Has-many-%3Athrough-association

https://trello.com/c/unNQd1fl