Send email along with cancellation events
Unfortunately, Segment doesn't send additional identify information via
webhooks, it just sends the raw information.

The Zapier Intercom integration only can create a message on users by email.

Therefore, we need to send the email along with this so Zapier can use it.

We may want to do something else in the future, but I think this is the best
solution for now.

https://trello.com/c/FIRl7bI2/539-collect-feedback-upon-cancelation