Remove coupon field from checkout form
The presence of the coupon field causes some customers to ask for a coupon,
cause hesitation when signing up, or to not sign up at all.
https://econsultancy.com/blog/7854-should-your-order-form-include-a-coupon-code-field

Now we can give users a coupon URL instead. When they follow this URL the
coupon will be placed in their session and automatically applied when they
checkout.

This change also simplies checkout by removing quantity. The user will no
longer select the quantity if they are subscribing to a team plan. It will
start them at the minimum quantity and then they can add people later.

This change also moves the minimum quantity of a plan out of a constant
into the plans table. This is more flexible and leads to cleaner code and
specs.

https://trello.com/c/h8lxKCTd/479-remove-visible-coupon-field-from-checkout-form