Remove legacy trails
Once upon a time, Upcase had these things called trails.

They broke down a topic into a series of steps that could be completed.
They also linked to resources (made by thoughtbot and not) that would
help you learn a topic.

We liked these trails, but the only way to indicate you'd learned
something was to click a checkbox on a given step.

Later, we created our exercise system. After creating a handful of
unrelated exercises, we decided it would be better to create groups of
exercises to be completed in a given sequence.

We decided to call this grouping and sequencing a trail and renamed the
old trail concept legacy trails.

Having two things called "trail" at the same time was confusing, and the
legacy trails were the source of much criticism ("these are mostly links
to things outside Upcase!") We didn't like the code. Users were confused
and annoyed. Bad times.

This commit removes LegacyTrails and their associated baggage.

Now, "trail" means exactly one thing.

All is :sparkles: and :joy:.