Update a user's card when they sign up again
This fixes a bug for users who sign up, cancel, and try to sign up
again.

Before this fix, we attempted to charge their *old*, original
card, even when they provided a new one for us to use. This led to
confusing errors like us reporting that someone's clearly-not-expired
card was.

Additionally, this commit removes the ability for users to sign up a
second time using the a card we have on file for them. That
functionality is not well-supported by our Stripe fake, and made the
completion of this bugfix substantially more difficult. Users who are
signing up a second time after canceling will have to enter their card
again. I'm comfortable removing this since it simplifies a very complex
part of the codebase, and the functionality lost benefits a small
portion of members.

https://trello.com/c/VljVeZS3/8