Add endpoint to redirect from whetstone back to trail
Currently users are redirected to the `/practice` page with no additional
guidance after completing an exercise in the exercises app. This change exposes
a new route and controller that will allow us to redirect users back to a trail
after they complete an exercise, providing continuity.

This relies on an exercise having **one**, not many trails. All data matched
this assumption, but the code was implemented using a `has_many` relationship.
Since both exercises and videos act as completeables for, `Video` was also
changed to use a `have_one` trail relationship.