Dump schema to capture update to format
Rails seems to have updated the schema dump format to include limits and this
is the first change to introduce a migration since we upgraded. This commit
just captures the updated dump with no actual changes to the underlying schema
such that future migrations will only produce a db/schema.rb diff related to
their change.