Add environment banner in staging and dev
The environment banner displays the environment name as well as the current
branch. This is useful to visually distinguish environments (including
production by the lack of a banner), as well as quickly answering the question
of what is currently deployed.

Unfortunately Heroku deploys only the code, not the git repo, so branch and
commit sha must be pushed from the deploying environment which is done via the
bin/deploy script.