Remove `public` attribute from Exercises
Previously it was expected that exercises could be displayed independently on a
Topic page, and the `public` flag would determine their visibility. Now all
exercises are expected to be viewed within the context of a trail and the
trail's `published` flag determines the visibility of both the trail and any
associated exercises.

With this in mind we are removing the `public` field to simplify the exercise
model.

Associated to this, we've determined that Exercises should not be classified
independently as they are intended to be viewed as part of a trail, so this
commit also removes the Topic association via Classifications for Exercises and
removes the classification data.