Fix permission checks for videos on trails
We never removed the assumption that all videos belong to a watchable,
even though videos belonging to a trail are associated through a step
instead.

Existing videos on trails both belongs to a step and a (hidden)
watchable.

Permissions checks were going through the watchable, even though users
could only actually access videos through the trail.

Because the previous migration deletes all the video tutorials, we need
to make sure videos can function with a step and no watchable.

https://trello.com/c/tjhwKYew/605