Display the github username field with errors
The Checkout form object had logic to hide the github username when the
associated user instance already had a github username. Unfortunately, in the
case of a user who had created an account via github auth, but was now
attempting to sign up with email and password using their github username, the
sign up would fail, and would re-render the sign up form without the github
username field, preventing the "Github username already taken" message from
being displayed.

This update adds a condition to the `#needs_github?` method on Checkout to
also display the field if there is an error associated with the github username
field.

The `semantic_errors` call was removed as it did not seem to produce any
output, and the expected error messages are displayed inline with the fields.

PR feedback

Rename to needs_github_username