Add support for test env and in bare pry repl
Assuming the presence of the `Rails` constant and the environment definition
caused pry to crash on startup if run directly as `pry` in the shell. This
fixes by only drawing the custom prompt if in a Rails context.

In addition, there was no color mapping for the "test" environment so pry would
crash when started from within a spec. This commit fixes that by explicitly
setting the colors for production and staging, and using the default reset
color for all other envs.