Revert Capybara.ignore_hidden setting to true
In a previous upgrade the `Capybara.ignore_hidden_elements` setting was
switched to `false` which has the effect of letting Capybara `find` elements on
the page that a user would not see. This looks to be associated to an upgrade
of Rspec and was likely used as a temporary fix to get the suite back to green.

This update reverts the configuration to once again have Capybara ignore hidden
elements. This more closely models a user interacting with the page and is thus
a good default. When needed, Capybara finders and matchers can use the
`visible: false` option to explicitly find hidden elements such as meta tags
and RSS feed links.