Add admin form for creating quizzes and questions
Previously we relied entirely on rails admin for our admin needs. While it
worked well for simple CRUD management, it quickly failed us as our workflows
become more complex such as trail & exercise management. Further, we've had a
number of instances where rails admin functionality has broken without us
realizing leaving us scrambling when we needed to update something.

This PR introduces a distinct admin interface for the quiz & questions flow.
We've implemented the main features we need and have test coverage to prevent
accidental breaking. It does not cover everything, but provides a
straightforward interface for the main workflows while assuming anything not
implemented can be handled in the console.

The specific functionality implement is:

- Listing quizzes
- Creating a new quiz
- Viewing questions for a quiz
- Adding / editing a question (with live preview)
- Re-ordering the questions within a quiz.