Instruct spring to watch additional routes files
We are defining our routes in multiple files that are required into the main
config/routes.rb file, but Spring is unaware of this. This means Spring needs
to be restarted to pick up changes not in the main routes file.

This commit updates the Spring configuration to also watch all files in the
`config/routes/` directory.