Show users questions on which they were not confident
When a user rates their confidence at a 3 or lower, add it to a list of
questions that need to be reviewed, displayed on the quiz index.

This makes it easy for users to review questions with which they
struggled. This is particularly nice because they can review supporting
material outside of the quiz session itself (and possibly on a
non-mobile device).

If a user re-answers a question and rates it a 4 or 5, it will be
removed from the review list.