Hide the nav on question and quiz pages on mobile
The nav links in the header take up a considerable amount of the vertical space
on mobile. On most pages this is fine, but on the quiz pages it means needing
to scroll down on every single question to see the content.

This change hides all navigation elements other than the primary "Upcase" link
when users are viewing a quiz, question, or the results page.