Add fastclick for better touch responsiveness
Mobile browsers by default wait a 300ms delay between registering a touch event
and firing the corresponding click event. This is to allow time to register a
double tap used to resize when viewing a site sized for desktop. Since we have
designed our site responsively, this is not needed and introduces an
unfortunate laggy feel to our UI.

Fastclick registers event listeners for touch events on mobile that immediately
fire the click event to remove the lag. No event listeners are added on desktop
or mobile browsers that don't add the delay (Chrome 32+ with proper viewport
meta tag).