Update the markdown parser and syntax highlighter
Previously we were using bluecloth for markdown parsing. This updates the app to
use redcarpet which can parse more closely to the Github markdown we tend to
author in. Specifically this gives us the ability to include tables and fenced
code blocks.

This also swaps highlight.js in for prettyify to more accurately highlight our
code blocks. This also works well with the fenced code blocks exposed by
redcarpet's renderer to specify the language to highlight.