Syntax highlight in the admin preview
The hljs initializer in application.js hooks into the page load event and as
such only fires once. Since the preview content is dynamically loaded in after
page load, and is re-populated on all changes, it is never highlighted.

This change hooks into the update event to trigger the highlighting after the
content is replaced.