Switch from RailsAdmin to Administrate
Administrate is a Rails engine that provides an easily customizable
Admin dashboard with a clean UI.

Administrate is heavily inspired by projects
like Rails Admin and ActiveAdmin,
but aims to provide a better user experience for site admins,
and to be easier for developers to customize.

To do that, we're following a few simple rules:

- No DSLs (domain-specific languages)
- Support the simplest use cases,
  and let the user override defaults
  with standard tools such as plain Rails controllers and views.
- Break up the library into core components and plugins,
  so each component stays small and easy to maintain.

See more at: https://administrate-docs.herokuapp.com/

Motivation
==========

Switching to Administrate gives us a chance to dogfood the library
before releasing it to the world as open-source.

Changes
=======

- Install Administrate & update dependencies
- Fix deprecation warnings from new SASS version
- Run `rails generate administrate:install`
- Remove Rails Admin
- Implement admin authorization
- Customize dashboards based on previous Rails Admin config
- Implement `#to_s` methods for models that were missing it
  (See https://trello.com/c/9OfUvC3x).

Regressions from RailsAdmin
===========================

- No built-in searching/filtering of records.
  This is fairly easy to add in for Upcase,
  and there is a Trello card in progress for it:
  https://trello.com/c/CnV9wB29
- No JSON/CSV/XML exporting.
  Again, this would be fairly easy to implement within Upcase,
  and there's a Trello card to add it in the future:
  https://trello.com/c/IIsGsngL

Known bugs for Upcase
=====================

- NameError - uninitialized constant Article
  - https://trello.com/c/h9DXLzkQ
  - `admin/topics#edit`
  - This was present in RailsAdmin as well
- Timeouts
  - https://trello.com/c/4HFXALFm
  - `admin/plans#edit`
  - `admin/teams#edit`
- ActiveRecord::AssociationTypeMismatch
  - https://trello.com/c/7sXTYFC3
  - `admin/exercises#update`
  - `admin/subscriptions#update`

Conclusion
==========

If any of these regressions or bugs are show-stoppers,
let me know and I can update Administrate to take care of them
before merging this in.

Otherwise,
the linked trello cards are queued up to be fixed in upcoming releases.