Add 'View In Video' buttons to seek specific times
The wistia player API exposes functionality to seek specific times in a given
video. This update allows "Markers" to be defined for a video which connect a
specific time in the video to a given topic (specified by the header ID
aligning with the notes).

Overall this allows for users to easily revisit specific parts of a video
without having to rewatch the whole thing.

Specific features:

- Generate anchor link ids when rendering markdown
- Render 'View Topic In Video' seek buttons
- Update the URL hash, scroll the video into view, and play at the marker on click
- Start playing (or unpause) on video button click