Instruct wistiaEmbed to prefer HTML5 player
Many users were getting the Flash player despite having a browser capable of
playing HTML5 video. The flash player has a number of limitations and drawbacks
when compared to the HTML5 player, notably in the player control API.

This sets an explicit preference for the HTML5 player when embedding, but will
still fallback to Flash in cases where HTML5 video is not supported.