Add Explore menu to signed-out header
Exposing all of the content from a mega-dropdown in the nav bar lets
potential users see both the volume of material and the quality of
topics. This could persuade potential users to sign up.

The body overflow is hidden when open to prevent inconsistent behavior
between the different "sticky" header states. It also avoids accidental
background scrolling.