Create a new checkout flow
The motivation here is to make signing up simpler, which will hopefully
convince more people to do it.

This commit does not remove the existing checkout flow, as we plan to
A/B test the two against each other.

Note that this checkout flow isn't linked to from anywhere, so can be
deployed even though there are placeholder pages.

Major changes in the new flow:

* Users *must* signup using GitHub OAuth.
* Credit card details are entered using Stripe Checkout, not our own
form.

Known issues:

* Users will not be able to *re*subscribe using this flow.
* The join page and the payment page are both placeholders.