Add thoughtbot branding to nav bar
One of the leading factors in signups was the strong reputation of the
thoughtbot brand. Adding a tagline and Ralph to the header could
reinforce the association, encouraging more signups.

We're only showing Ralph when the hero isn't visible because the hero
will have also have Ralph in its CTA. That would add tension and
redundancy.