Remove the recursion limit on StripeCustomerCollection
This appears to exist just to prevent a "stack level too deep" issue if
someone used a very small per_page limit and we had thousands of
customers and not very much memory.

Currently, the only code that uses this uses a per_page value of 100, so
I think the risk is low.

This also happens to delete our costliest test, in terms of time (3
seconds) since we create hundreds of uuids and do lots of iteration.