Display viewed status for Weekly Iteration videos
We've always collected the statuses as users watch the Weekly Iteration on the
site using the same AJAX calls the statuses endpoint, but up till now this was
hidden.

With this PR, we expose the watched state via the icon, as well as exposing the
"Mark as Complete" button below the video itself.