Clarify distinction between canceling and deactivating subscriptions
When a user wishes to cancel her subscription, there are two distinct
events:

1. The user clicks cancel in the app. They can continue to access all
content until their current billing cycle ends.
2. Their billing cycle ends, and they can no longer access any content.

Currently, we are not tracking the date at which #1 occurs.

Subscription has a scheduled_for_cancellation_on field, which tracks
that date at which #2 will happen. We use this to tell the user "you can
keep using Upcase until #{date}".

It also has a deactivated_on field, which tracks when #2 actually
happens.

These two fields will be the same for nearly all users. (The exception
is if we refund a user, their account is deactivated immediately. We
also deactivate immediately if we move them to a team plan.)

There are two problems with the current approach.

1. We are using inconsistent language to refer to the same thing:
"cancel" vs. "deactivate".

2. We aren't tracking the date at which the user decides to cancel and
clicks the cancel button. (We can technically determine this data from
the date we send the "Cancel" event to analytics, but I'd like to have
this data in our db as well.")

This commit addresses both issues.

First, we now consistently refer to the loss of content access as
deactivation.

Second, we now store the date a user clicked cancel in a
user_clicked_cancel_button_on field.

And all was right with the world.