Use timezone-aware methods in cancellation spec
The date under test is stored in the application timezone, UTC. The
test, however, used `Time.now` and `Date.today`, both of which use the
system timezone. This cause a comparison between my system timezone
(EDT) and the application timezone (UTC) resulting in the test failure.

This PR changes `Time.now` and `Date.today` to user their application
timezone equivalents, `Time.current` and `Date.current`. This passes
the test by ensuring the date comparison is UTC to UTC.