Fix bug where Start Trail would display wrong step
Before this commit, clicking Start Trail on some trails would load a
step that wasn't actually the first step of the trail.

This is because the `by_topic` scope was sorting the results by topic
name, and overriding the ORDER BY on steps.position from acts_as_list.

Now, we explictly add the steps.position ordering in a
first_completeable method on Trail, which ensures this ordering is
retained.

Additionally, this commit extracts a TrailsForPracticePageQuery and
injects decorated trails directly into Practice. This simplifies the
crap out of the tests for Practice and probably should have been done
long ago. I am, however, at peace with my failure to do so earlier.
Namaste.