Provide email to Stripe for GitHub authed users
If the user elects to authenticate with GitHub during checkout, then
they do not fill in the "email" field on the `Checkout` form object. We
were previously using this field when creating the customer object in
stripe, which causes the email to blank in Stripe.

All checkout objects have an associated user, so we can use the
previously-existing delegation `user_email` rather than the virtual
field that exists for the form only.