Improve performance of flashcards needing review
Previously this query iterated over every flashcard in the system as a
ruby object, queried the database for the most recent attempt for the
current user, then iterated over those resulting attempts to select
those that were below a confidence threshhold before finally mapping
over that set once more to get the flashcard object back.

The `latest_attempts` view added here presents a list of attempts that
are only the most recent for each user and flashcard. We can then query
that without the use of ruby. This eliminates a whole mess of queries
and Ruby object allocations.