Eliminate N+1 query for video status
Each video shows a completion status for the current user. This was
previously an N+1 query. By adding the `ShowListing` decorator, we can
expose a collection of videos that the `ShowListing` can decorate with
the proper status, having done a single query for all of the statuses
for each video in the show.