Remove the Change Plan link
This is part of a larger process to simplify the plan structure such that all
subscribers have the same access. Eventually we want to remove the edit
subscription page entirely, but that would require removing the Weekly
Iteration plan and all the associated support code.

This was the simplest step we could take to remove the "Change Plan" link which
should prevent new upgrades / downgrades.