Inline the routes
Breaking the routes out into separate files was a worthy experiment but
it seems the team feels it's more a bother than anything else. Finding
where a given route is defined or choosing where to put a new route is
more trouble than it feels like it should be.

Additionally, the tucking away of routes is hiding some unnecessary
performance overhead that we can avoid by re-ordering, combining, or
improving the routes.