Move `root` route to the top of the routes file
Previously, `root` was situated below dozens of other routes, which
means the router will attempt to match all of those routes first. We
know that for a root url, none of those will match. Further, some of the
routes that are ahead of it, use constraints that issue database
queries.

Moving `root` to the top of the file avoids 3 unnecessary SQL queries
and several dozens of route matches. This improved root URL response by
60-70ms in my local testing.