Improve SlugConstraint performance
The SlugConstraint is used three times within the routes. Before this
change, each time it was used it would issue a separate database query.
If the route you were actually hitting was listed below the routes that
use the constraint you'd be hitting 3 needless database queries before
you even reached your route.

This change ensures the routes will issue a maximum of 1 database query,
no matter where the eventual route match is. It:

1. Introduces a `slugs` database view that is a single table that lists
   all slugs and their associated model.
2. Updates `SlugConstraint` to look up slugs by id in this view and
   compare the `model` column to the model class that was passed in to
   determine a match.
3. Caches the result of the slug type lookup in `[RequestStore][1]` so
   that the following uses of SlugConstraint don't have to perform that
   query again and can just do a string comparison.

[1]:https://github.com/steveklabnik/request_store