Enhance trail show view for non-subscribers
This updates the trail display to emphasize the teachers as well as display a
title card if one exists.

In addition, this change simply lists the titles in a trail rather than
providing links, since those links would confusingly take users to the sign in
page with the message "You must be the subscription owner"

This is in preparation for the changes in work to allow free access to sample
videos in the trails.