Use successfully tested new landing page for all
The A/B test proved the new landing page to be an improvement over the
existing landing page, so we're baking in that change. I removed
associated tracking calls and unused vanity methods as well.

At Ben's request, I left the vanity gem installed and the routes to its
engine running as this will ease setting up any future A/B tests.

I *did* remove the `use_vanity` call in `ApplicationController`. This
method causes database queries in order to look up the current user's
A/B test settings on every request. In future A/B tests, we should add
this to impacted controllers only to limit the scope.