Update rss format, remove /videos.rss
In hopes of providing a more useful and readable RSS summary, this commit
updates the rss builder code to use the rendered markdown of the summary of a
video for the rss description.

This also removes the /videos.rss feed which seemingly was meant to serve all
the various "shows", but in the end we only have the Weekly Iteration so this
feed is largely a duplicate of the Weekly Iteration feed.