Update to RSpec 3.3
This will enable us to use niceities such as `bisect` and re-run failed
examples or just the next failure. Two changes needed to be made for 3.3
compatibility:

First, I had to make the `Trail.completed_for` test match an array of
ids rather than array of  objects due to a change in how rspec 3.3
matches objects that have been decorated. See:
https://github.com/rspec/rspec-core/issues/1995

Second, I had to make make `raise_error` assertions specific.  A vanilla
`raise_error` assertion can catch any number of errors, including
NoMethodError (the thing you are testing isn't actually a method),
errors with your mocks, etc. Rspec 3.3 rightfully warns on these.

The tests removed from the cancellation spec were super difficult to
follow and once I got them working I realized that they were not testing
what they claimed to be testing. That is, they claimed to test that an
error from ActiveRecord in the cancellation process would cause the
stripe subscription to not be canceled. Unfortunately, the stripe
cancellation call is a web service call and is not subject to
transaction rollback. There were no assertions that the stripe
subscription wasn't canceled, thus the tests *could* pass.

The cancellation code and specs could use another look to be sure we
have adequate coverage (and the specs should rely less on doubles and
essentially global state) but the scope of this change was to clean up
`raise_error` warnings, so I opted to remove the tests that aren't
testing what they claim to be.