Replace multiple layouts with visitor & signed_in
Previously there were five distinct layouts, with some overlap between them.
This change simplifies down to just two: signed_in and visitor, giving us
consistency and focus on these two main scenarios.

In working on this we realized that there is also a third potential state of a
signed_in user who does **not** have an active subscription, but for the time
being we are just treating them as a signed_in user and will revisit this as we
add support for previews and the like.