Force SSL via Rails config
We were previously using `Rack::SslEnforcer` but this has been
superseded by functionality built into rails (`config.force_ssl`) which
is based on `Rack::SSL`. This does a little bit more than the previous
middleware in that it also enables HSTS, which means that subsequent
protocol-less visits to the domain will default to HTTPS.

`Rack::SslEnforcer` had a canonical domain redirect feature built in.
We'll use `Rack::CanonicalHost` for this, which brings us in line with
the typical setup from suspenders these days.

Also removed the rack-rewrite code that was doing a canonical host
redirect as `Rack::CanonicalHost` is higher in the middleware stack.