Move redirects from routes to Rack::Rewrite
By having these in routes, they all generate path and url helper
methods, pollute `rake::routes` and generally cause some non-negligable
cognitive overhead when looking through routes.

There were a few places where we were actually using the url helpers in
code, which results in a less than ideal situation of users paying the
performance cost for redirects when we could just as easily send them
where we want them to go.

The product redirections with regex constraints that were in the routes
file were overly greedy. We were attempting, for instance to match
products ids 10 or 12 but were also matching any product ids that
started with 10 or 12. These were fixed in the migration to rewrite.